backend/requirements.txt
fastapi
uvicorn
motor
pydantic-settings
python-jose
passlib[bcrypt]
email-validator
